Adding my current code
This is the code for the html, css files i currently have for the webapp. I had to remake it because I could not figure out a bug regarding the password hashing. This code is connected to an sql database called pythonlogin2. Users can create an account, then log in and have access to the homepage (currently empty), a profile tab that shows their infromation, and a log out button that will end their session and return the user to the login page.